fix warnings

+++ drivers/spiflash.c
@@ -105,7 +105,7 @@ static uint32_t get_num_bytes(void) {
 
 void spiflash_dump_sfdp(void) {
     uint64_t id = spiflash_unique_id();
-    DMESG("flash ID %x %x", (uint32_t)(id), (uint32_t)(id >> 32));
+    DMESG("flash ID %x %x", (unsigned)(id), (unsigned)(id >> 32));
 
     sfdp_header0_t hd0;
     spiflash_read_sfdp(0, &hd0, sizeof(hd0));
@@ -118,11 +118,11 @@ void spiflash_dump_sfdp(void) {
         sfdp_header_t hd;
         spiflash_read_sfdp(sizeof(hd0) + sizeof(hd) * i, &hd, sizeof(hd));
         uint32_t addr = (hd.pointer[2] << 16) | (hd.pointer[1] << 8) | hd.pointer[0];
-        DMESG("SFDP: header %d id=%d ptr=%d len=%d", i, hd.param_id, addr, hd.length);
+        DMESG("SFDP: header %d id=%d ptr=%d len=%d", i, hd.param_id, (int)addr, hd.length);
         for (int j = 0; j < hd.length; ++j) {
             uint32_t v;
             spiflash_read_sfdp(addr + j * 4, &v, sizeof(v));
-            DMESG("  %x", v);
+            DMESG("  %x", (unsigned)v);
         }
     }
 }
@@ -196,7 +196,7 @@ void spiflash_init(void) {
 
     _spiflash_num_bytes = get_num_bytes();
 
-    DMESG("SPI flash %d Mbytes", (_spiflash_num_bytes + 511 * 1024) >> 20);
+    DMESG("SPI flash %d Mbytes", (int)((_spiflash_num_bytes + 511 * 1024) >> 20));
 }
 
 #endif
